Ensure that all series values are floats
There are instances I've run into where zeppelin (0.9.0-preview1) may interpret your numeric data as strings, **regardless of whether or not they are actually float values underneath**. This is unfortunate, as it also leads to [Highcharts Error 14](https://www.highcharts.com/errors/14/) trying to access the inner series in the chart. Putting in a `parseFloat` call here is one easy way to ensure the data are numeric when passing into the HighCharts instance.